+++ abserv/abserv/Npc.h
@@ -48,7 +48,7 @@ public:
 
     void Update(uint32_t timeElapsed, Net::NetworkMessage& message) override;
     bool Die() override;
-    bool Resurrect(int16_t precentHealth, int16_t percentEnergy) override;
+    bool Resurrect(int precentHealth, int percentEnergy) override;
     std::shared_ptr<ai::AI> GetAi();
     void Shutdown();
